Release screen shares the lockup offer's ReviewPlanCard layout
Both are the same "commit to some reviews" moment, so they now render
identically: title, cards grouped by type, one button, week strip.
ReleaseOffer gains a release_preview (CardSummary list) on the Rust
side; WeekProgressStrip moves to its own file to avoid a circular
import between LockupOffer and NoCardsReady.
